Render cell tower location and coverage on map

+++ app/src/main/java/com/tower/locator/model/LocateResponse.kt
@@ -1,8 +1,8 @@
 package com.tower.locator.model
 
 data class LocateResponse(
-    val lat : Int?,
-    val lon : Int?,
-    val radius : Int?,
+    val lat : Double?,
+    val lon : Double?,
+    val radius : Double?,
     val error : String?
 )
